+++ app/Http/Controllers/UserController.php
@@ -6,8 +6,8 @@ use App\CustomClasses\MailList\MailListFacade;
 use App\Exports\OldUsersExport;
 use App\Exports\PennoMemberExport;
 use App\Exports\UsersExport;
+use App\Models\Rol;
 use App\Repositories\UserRepository;
-use App\Rol;
 use App\Rules\EmailDomainValidator;
 use App\User;
 use Illuminate\Http\Request;
@@ -15,9 +15,6 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Session;
 use Maatwebsite\Excel\Facades\Excel;
-use App\Events\MemberKindChanged;
-use App\Events\MemberBecameOldMember;
-use App\Events\OldMemberBecameMember;
 
 class UserController extends Controller
 {
