+++ database/factories/RoleFactory.php
@@ -2,7 +2,7 @@
 
 namespace Database\Factories;
 
-use App\Rol;
+use App\Models\Rol;
 use Faker\Generator as Faker;
 
 $factory->define(Rol::class, function (Faker $faker) {
