+++ database/seeders/ClearSeeder.php
@@ -11,8 +11,8 @@ use App\Models\Book;
 use App\Models\Certificate;
 use App\Models\MenuItem;
 use App\Models\NewsItem;
+use App\Models\Rol;
 use App\Models\UserEventLogEntry;
-use App\Rol;
 use App\User;
 use App\Zekering;
 use Exception;
